+++ backend/__init__.py
@@ -1,4 +1,4 @@
 """Package level information
 """
 
-__version__ = "5.4.0"
+__version__ = "6.0.0"