+++ frontend/src/containers/TotalFlashcardBrowser/index.js
@@ -1,1 +1,2 @@
-export {default} from './TotalFlashcardBrowser';
+export {default as TotalFlashcardBrowser} from './TotalFlashcardBrowser';
+export {default as PreviewTotalFlashcardBrowser} from './PreviewTotalFlashcardBrowser';